Remove null mask and null count from column_view constructors (#13311)
This is a breaking change that removes default values. Removing the default null count value of `UNKNOWN_NULL_COUNT` is necessary since we are removing `UNKNOWN_NULL_COUNT` altogether. A potential alternative was setting the default to 0, which would correspond to a default null mask of `nullptr` (the current default). However, that change is potentially more dangerous if callers were previously setting the null mask to something other than `nullptr` without setting the null count and relying on the behavior of `UNKNOWN_NULL_COUNT`. Removing the default parameter altogether is therefore the safer option here.

Contributes to #11968.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Nghia Truong (https://github.com/ttnghia)
  - Vukasin Milovanovic (https://github.com/vuule)

URL: https://github.com/rapidsai/cudf/pull/13311