Updates needed to work with jitify that leverages libcudacxx (#13383)
This update libcudf with a version of jitify that leverages libcudacxx. This removes compiler issues when the pre-built defines in jitify and libcudacxx diverege due to updates in either package.

At the same time testing locally with CUDA 11.5, 11.8, 12.0, and 12.1 uncovered a couple of compile errors.

Authors:
  - Robert Maynard (https://github.com/robertmaynard)

Approvers:
  - Nghia Truong (https://github.com/ttnghia)
  - Vukasin Milovanovic (https://github.com/vuule)

URL: https://github.com/rapidsai/cudf/pull/13383