Update JNI JSON reader column compatability for Spark (#13477)
This moves the logic to update the columns returned from the JSON reader to java. It also updated the code to be able to deal with requested columns that were not in the data. It is not perfect because it will not work if the input file had no columns at all in it.

```
{}
{}
```

But it fixes issues for a file that has valid columns in it, but none of them are the columns that we requested.

This is a work around for https://github.com/rapidsai/cudf/issues/13473, but is not perfect.

Authors:
  - Robert (Bobby) Evans (https://github.com/revans2)

Approvers:
  - Jason Lowe (https://github.com/jlowe)
  - MithunR (https://github.com/mythrocks)

URL: https://github.com/rapidsai/cudf/pull/13477