Simplify implementation of interval_range() and fix behaviour for floating `freq` (#13844)
Closes #13843 
Closes https://github.com/rapidsai/cudf/issues/13847

This PR simplifies the implementation of `interval_range()` and fixes a few different bugs in the process. It also moves all tests for interval indexes to `tests/indexes/test_interval.py`. Finally, while working on this PR, I ran into #13847; a fix for that is also included in this PR.

Authors:
  - Ashwin Srinath (https://github.com/shwina)

Approvers:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/13844