Correctly detect the BOM mark in `read_csv` with compressed input (#13881)
We currently detect the BOM mark at the start of the file and then decompress the data, if compressed. This leads to a problem with compressed files that have a BOM mark, as we are trying to detect the mark in the compressed data.
This PR moves the BOM detection until after decompression. Also cleaned up the surrounding code slightly.

Authors:
  - Vukasin Milovanovic (https://github.com/vuule)

Approvers:
  - David Wendt (https://github.com/davidwendt)
  - Vyas Ramasubramani (https://github.com/vyasr)

URL: https://github.com/rapidsai/cudf/pull/13881