Fix an issue where casting null-array to `object` dtype will result in a failure (#13994)
closes #13992 

This PR fixes the construction of an empty `MultiIndex` from `pandas` to `cudf` was causing an error in pandas-compatibility mode, null-array is one such case where it is _okay_ to cast to any type because there is no data in it. Hence we pass `str` dtype to `astype`, whenever we encounter an `object` dtype.

Authors:
  - GALI PREM SAGAR (https://github.com/galipremsagar)

Approvers:
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cudf/pull/13994