Fetch correct `branch-23.10-xdf` in `-main` jobs (#33)
We currently have -main related jobs to build wheels and compare the current PRs pandas tests with branch-23.10-xdf. In these -main jobs, we perform git checkout branch-23.10-xdf, this will actually fetch the latest branch-23.10-xdf from rapidsai/cudf-private, which is most likely to be up-to-date than when the PR's branch has been created. This is causing the pandas test reporting to compare upstream/branch-23.10-xdf diff to the current PR which basically gives us misleading results unless we update the current branch again with upstream/branch-23.10-xdf. Thus to correctly compare just the PR's impact we need to count and revert the number of commits a PR is introducing.